refactor: split cache invalidation by domain
